src/datasets/cityscapes/requirements.txt
pillow
tqdm
cityscapesScripts==1.5.0
